Wire context-menu "Read with Zeeguu" to the new upload flow
The context-menu handler still injected the old contentScript.js
(InjectedReaderApp), so right-click → "Read with Zeeguu" was broken.

- Extract scrape + POST /article_upload/create into a shared
  sendTabToZeeguu helper used by both popup and background.
- Background's startReader now opens /shared-article?upload_id=<id>
  in a new tab, consistent with the popup path.
